requirements.txt
opencv-python-headless
requests
asyncio
lumaai
python-dotenv
openai
